% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{add_IMD}
\alias{add_IMD}
\title{add IMD (boundary change https://www.gov.uk/government/statistics/2011-rural-urban-classification-lookup-tables-for-all-geographies: Rural Urban Classification 2011 lookup tables for local authority areas)}
\usage{
add_IMD(dat)
}
\description{
add IMD (boundary change https://www.gov.uk/government/statistics/2011-rural-urban-classification-lookup-tables-for-all-geographies: Rural Urban Classification 2011 lookup tables for local authority areas)
}
