% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4paper.R
\name{add_legend_main}
\alias{add_legend_main}
\title{Add a title to the legend of the correlation plot}
\usage{
add_legend_main(x, y, main)
}
\description{
Add a title to the legend of the correlation plot
}
