% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4paper.R
\name{add_one_forecast}
\alias{add_one_forecast}
\title{used for producing the wastewater vs wastewater+national-prevalence nowcast}
\usage{
add_one_forecast(f, cols)
}
\description{
used for producing the wastewater vs wastewater+national-prevalence nowcast
}
