% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{add_tmajd_to_fitdata}
\alias{add_tmajd_to_fitdata}
\title{Add RW neighbourhood structure to fitdata}
\usage{
add_tmajd_to_fitdata(fitdata, nts, RW.order = 1)
}
\description{
Add RW neighbourhood structure to fitdata
}
