% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions4paper.R
\name{compare_forecasts}
\alias{compare_forecasts}
\title{used for producing the wastewater vs wastewater+national-prevalence nowcast}
\usage{
compare_forecasts(fs, obs, setting)
}
\description{
used for producing the wastewater vs wastewater+national-prevalence nowcast
}
