% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expit}
\alias{expit}
\title{expit transform (the inverse of logit)}
\usage{
expit(a)
}
\arguments{
\item{a}{a logit value to be back-transformed}
}
\value{
expit(a)=exp(a)/(exp(a)+1)
}
\description{
expit transform (the inverse of logit)
}
