% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{extract_all_parameters}
\alias{extract_all_parameters}
\title{a wrapper function of extract_params to extract a collection of parameters}
\usage{
extract_all_parameters(params, output, start, end = NULL, ndraws = NULL)
}
\description{
a wrapper function of extract_params to extract a collection of parameters
}
