% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{extract_from_alldata}
\alias{extract_from_alldata}
\title{Extract prevalence and wastewater data from the entire set for a specific model fit}
\usage{
extract_from_alldata(imodel, fit_ts, horizon, idraw, alldata, covars = NULL)
}
\description{
Extract prevalence and wastewater data from the entire set for a specific model fit
}
