% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{extract_params}
\alias{extract_params}
\title{extract a given parameter from nimble fit}
\usage{
extract_params(param, dms, output, start, end = NULL, nms)
}
\description{
extract a given parameter from nimble fit
}
