% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{forecast_summary}
\alias{forecast_summary}
\title{summarise quality of disaggregated prevalence}
\usage{
forecast_summary(pred_mean, pred_ci, obs)
}
\description{
summarise quality of disaggregated prevalence
}
