% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{format_ltla_prevalence_wide_sims}
\alias{format_ltla_prevalence_wide_sims}
\title{Import weekly wastewater estimates for the whole of England}
\usage{
format_ltla_prevalence_wide_sims(
  ltla_prev,
  nsims,
  nts_prev,
  ltla_ids_ww,
  weekly_ids_ww,
  get_sims = TRUE
)
}
\arguments{
\item{nts}{Number of weeks}
}
\description{
This function loads the weekly wastewater estimates for the whole of England. These are population weighted average of the
LSOA-level weekly wastewater estimates
}
