% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_all_ww_prev_data}
\alias{get_all_ww_prev_data}
\title{Get all LTLA/national weekly prevalence and WW data}
\usage{
get_all_ww_prev_data(save_file = NULL, load_file = NULL)
}
\description{
Get all LTLA/national weekly prevalence and WW data
}
