% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.R
\name{get_inits_and_params}
\alias{get_inits_and_params}
\title{Specify initial values and parameters to save}
\usage{
get_inits_and_params(imodel, fitdata)
}
\description{
Specify initial values and parameters to save
}
