% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_subset.R
\name{get_inits_and_params_subset}
\alias{get_inits_and_params_subset}
\title{Specify initial values and parameters to save}
\usage{
get_inits_and_params_subset(imodel, fitdata)
}
\description{
Specify initial values and parameters to save
}
