% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_LTLA_population}
\alias{get_LTLA_population}
\title{Read in LTLA level population}
\usage{
get_LTLA_population()
}
\description{
data source: https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland
}
