% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_ltla_weekly_debiased_prevalence}
\alias{get_ltla_weekly_debiased_prevalence}
\title{Import LTLA-weekly debiased prevalence}
\usage{
get_ltla_weekly_debiased_prevalence()
}
\arguments{
\item{file}{Path to the data file}
}
\value{
debiased prevalence
}
\description{
This function loads the LTLA-weekly debiased prevalence
}
