% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_ltla_weekly_ww}
\alias{get_ltla_weekly_ww}
\title{Import LTLA-weekly wastewater estimates}
\usage{
get_ltla_weekly_ww(
  nts,
  ww_data_dir,
  ww_model = "ww_model_fit-20220801-195150",
  remove_2LTLAs = TRUE
)
}
\arguments{
\item{nts}{Number of weeks}
}
\description{
This function loads the LTLA-weekly wastewater estimates
}
