% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_national_prevalence}
\alias{get_national_prevalence}
\title{Obtain weekly national prevalence by taking population-weighted average of the LTLA-weekly prevalence}
\usage{
get_national_prevalence(prev_wide)
}
\description{
Obtain weekly national prevalence by taking population-weighted average of the LTLA-weekly prevalence
}
