% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_national_WW}
\alias{get_national_WW}
\title{Import weekly wastewater estimates for the whole of England}
\usage{
get_national_WW(nts, ww_data_dir, ww_model = "ww_model_fit-20220801-195150")
}
\arguments{
\item{nts}{Number of weeks}
}
\description{
This function loads the weekly wastewater estimates for the whole of England. These are population weighted average of the
LSOA-level weekly wastewater estimates
}
