% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{get_nimble_files}
\alias{get_nimble_files}
\title{get the file names for all fits associated with a given model and horizon}
\usage{
get_nimble_files(setting)
}
\description{
get the file names for all fits associated with a given model and horizon
}
