% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{get_nimble_MCMC_draws}
\alias{get_nimble_MCMC_draws}
\title{extract MCMC draws from a Nimble fit}
\usage{
get_nimble_MCMC_draws(imodel, draw_file, return_type, save_sims_list = TRUE)
}
\description{
extract MCMC draws from a Nimble fit
}
