% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_regional_WW}
\alias{get_regional_WW}
\title{Import weekly wastewater estimates for the whole of England}
\usage{
get_regional_WW()
}
\arguments{
\item{nts}{Number of weeks}
}
\description{
This function loads the weekly wastewater estimates for the whole of England. These are population weighted average of the
LSOA-level weekly wastewater estimates
}
