% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{get_weekly_LTLA_vax}
\alias{get_weekly_LTLA_vax}
\title{add LTLA-daily vaccine data}
\usage{
get_weekly_LTLA_vax(
  dat,
  ddir = "/Volumes/WorkSpace/onGitHub/wwprev/forTesting/vax/"
)
}
\description{
add LTLA-daily vaccine data
}
