% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{my_print}
\alias{my_print}
\title{Print a message on screen with a time stamp}
\usage{
my_print(m)
}
\arguments{
\item{m}{a message to print on screen}
}
\description{
Print a message on screen with a time stamp
}
