% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_subset.R
\name{nimble_models_subset}
\alias{nimble_models_subset}
\title{Specification of models in Nimble with spatially-coarse prevalence data only available for a subset of weeks during nowcast}
\usage{
nimble_models_subset(imodel, fitdata)
}
\description{
Specification of models in Nimble with spatially-coarse prevalence data only available for a subset of weeks during nowcast
}
