% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{point_in_interval}
\alias{point_in_interval}
\title{to check if a point (pt) is inside a given interval (itn)}
\usage{
point_in_interval(pt, itn)
}
\description{
Is itn\link{1}<= pt <= itn\link{2} true?
}
