% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{prepare_for_summary}
\alias{prepare_for_summary}
\title{preparing for nowcast summary (setting up working directory, file to save etc)}
\usage{
prepare_for_summary(setting)
}
\description{
preparing for nowcast summary (setting up working directory, file to save etc)
}
