% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{prepare_to_disaggregate}
\alias{prepare_to_disaggregate}
\title{carry out the preparation steps (specifying working directory, creating folders etc) for disaggregation}
\usage{
prepare_to_disaggregate(setting)
}
\description{
carry out the preparation steps (specifying working directory, creating folders etc) for disaggregation
}
