% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNimble.R
\name{save_nimble_fit}
\alias{save_nimble_fit}
\title{Save the nimble fit to file}
\usage{
save_nimble_fit(imodel, output, save_dir, fitdata, iters, region = NULL)
}
\description{
Save the nimble fit to file
}
