% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{sims_noise}
\alias{sims_noise}
\title{simulate a set of values from a normal distribution with mean 0 and given variance}
\usage{
sims_noise(s)
}
\description{
simulate a set of values from a normal distribution with mean 0 and given variance
}
