% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{summarise_dis_prev}
\alias{summarise_dis_prev}
\title{summarising disaggregated prevalence from multiple fits}
\usage{
summarise_dis_prev(ds, setting)
}
\description{
summarising disaggregated prevalence from multiple fits
}
