% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregation.R
\name{disaggregate}
\alias{disaggregate}
\title{disaggregating national prevalence to LTLA}
\usage{
disaggregate(sims_list_file, imodel, alldata, region = NULL)
}
\description{
disaggregating national prevalence to LTLA
}
